the database was getting wiped everytime time the code ran. put this as a variable with option to do it or not to retain data

+++ server/db.js
@@ -3,9 +3,10 @@ import { promisify } from "util";
 import fs from "fs";
 
 const DB_FILE = "grades.db";
+const RESET_DB = false;
 
 // Delete corrupted database if it exists
-if (fs.existsSync(DB_FILE)) {
+if (RESET_DB && fs.existsSync(DB_FILE)) {
   fs.unlinkSync(DB_FILE);
 }
